backend/requirements.txt
livekit
livekit-agents[silero,openai]
python-dotenv~=1.0
